cmake: don't set CL_INCLUDE_DIR in CMakeVendor.txt
This will always override anything the user provides. We could
turn it into a cache variable to avoid this but the default
value assumes the location of headers in a hypothetical driver
tree which probably doesn't work for anybody so just get rid of
it instead.

Also remove all references to OPENCL_DRIVER in the build system.
Any such integration doesn't really belong in the CTS's build
system and can triavialy be done with a wrapper project.

Signed-off-by: Kevin Petit <kevin.petit@arm.com>