Introduces the ability to capture the kernel sources and the build options used by the conformance tests to a .cl and .options files. This is achieved via a new compilation cache mode called dump-cl-files. Files will be generated to the folder pointed by the compilation cache path.
When a conformance test is executed with dump-cl-files cache mode and online
compilation mode, the files will be saved to the disk and the regular execution
flow of the conformance continues uninterrupted.

When a conformance test is executed with dump-cl-files cache mode and binary or spir-v
compilation mode, the files will be saved to the disk as well, however the compilation
will fail - leading to a failure in all test cases.