Framework to manage features
This change introduces a basic framework that makes it easier
to predicate the execution of a given test on a pre-defined feature.

Features are objects and reusable features can be defined as singleton
global objects (hence the need to initialise the test lists within the
main function of individual suites to guarantee that features are
already initialised).
Each feature comprises a description and a predicate function defined
as a member lambda. The predicate function returns true when the device
passed as argument supports the feature described by the containing
feature object.

Features can be used within expressions, either to define other features
or as part of test registration. Here are a couple of examples:

    const feature& F_complex_feature = F_featA && F_featB

    ADD_TEST_FEATURE(mytest, F_extension("cl_my_extension") || F_featA)

AND expressions stop evaluation as soon as one of the required features
is missing. This makes it possible to describe expressions that rely
on queries that are not supported by all versions of OpenCL.

The harness now skips tests based on whether they support the features
declared at registration time. It prints a message that describes the
feature expression that was evaluated when skipping a test. Here's an
example:

    progvar_prog_scope_init skipped (requires feature
    '((Version >= 2.0 && version < 3.0) ||
     (Version >= 3.0 && Program scope variables core optional))'
    that the device doesn't support

The uniqueness of the reusable objects make it possible to introduce a
caching mechanism to avoid repeating complex sequences of devices queries
but this change does not introduce such a mechanism.

This change uses the new framework to manage optionality in the api
and basic suites.

Fixes #792

Signed-off-by: Kévin Petit <kpet@free.fr>