Vulkan: Use linear tiling
The external memory sharing spec does not require
OpenCL to deduce if a Vulkan image is tiled or linear.
Use linear images by default and add a parameter
to test optimal images.

Additionally, when selecting a queue family, the Vulkan
implementation is not required to report the transfer bit
for compute or graphics queues, it is implied.